+++ config/config.exs
@@ -29,6 +29,7 @@ use Mix.Config
 #
 #     import_config "#{Mix.env}.exs"
 
+
 config :furlex, Furlex.Parser.Facebook,
   tags: ~w(
     fb:app_id fb:pages