+++ lib/furlex/parser.ex
@@ -20,6 +20,7 @@ defmodule Furlex.Parser do
     |> Stream.map(&extract(&1, html, match))
     |> Enum.reject(fn {_, v} -> is_nil(v) end)
     |> Map.new()
+    |> group_keys()
   end
   def extract(tag, html, match) do
     case Floki.find(html, match.(tag)) do
@@ -46,6 +47,49 @@ defmodule Furlex.Parser do
         |> Floki.attribute("href")
         |> Enum.at(0)
     end
+  end
+
+  @doc """
+  Groups colon-separated keys into dynamic map structures
+
+  ## Example
+
+  iex> Furlex.Parser.group_keys %{"twitter:app:id" => 123, "twitter:app:name" => "YouTube"}
+  %{
+    "twitter" => %{
+      "app" => %{
+        "id" => 123,
+        "name" => "YouTube"
+      }
+    }
+  }
+  """
+  def group_keys(map) do
+    Enum.reduce map, %{}, fn
+      {_, v}, _acc when is_map(v) -> group_keys(v)
+      {k, v}, acc                 -> do_group_keys(k, v, acc)
+    end
+  end
+
+  defp do_group_keys(key, value, acc) do
+    [h | t] = key |> String.split(":") |> Enum.reverse()
+    base    = Map.new [{h, value}]
+    result  = Enum.reduce t, base, fn key, sub_acc ->
+      Map.new([{key, sub_acc}])
+    end
+
+    deep_merge(acc, result)
+  end
+
+  defp deep_merge(left, right) do
+    Map.merge(left, right, &deep_resolve/3)
+  end
+
+  defp deep_resolve(_key, left = %{}, right = %{}) do
+    deep_merge(left, right)
+  end
+  defp deep_resolve(_key, _left, right) do
+    right
   end
 
   defp do_extract_content(elements) do