+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v.0.1.2
+* Parse colon-separated OpenGraph and Twitter Card keywords into map structures
+  - E.g. %{"twitter:app:id" => 123} becomes %{"twitter" => %{"app" => %{"id" => 123}}}
+
 ## v.0.1.1
 * Fix test-breaking bug when included as hex dependency
 * Add tags